% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_model.R
\name{ei_model}
\alias{ei_model}
\alias{ei_model.formula}
\alias{ei_model.ei_spec}
\alias{ei_model.data.frame}
\alias{ei_model.matrix}
\alias{ei_model.default}
\title{Fit King's ecological inference model}
\usage{
ei_model(x, ...)

\method{ei_model}{formula}(formula, data, weights, bounds = NULL, ...)

\method{ei_model}{ei_spec}(x, bounds = NULL, ...)

\method{ei_model}{data.frame}(x, y, z, weights, bounds = NULL, ...)

\method{ei_model}{matrix}(x, y, z, weights, bounds = NULL, ...)

\method{ei_model}{default}(x, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item An \link{ei_spec} object containing the outcome, predictor, and covariates.
}

Predictors must be proportions that sum to 1 across rows.
You can use \code{\link[=ei_proportions]{ei_proportions()}} to assist in preparing predictor variables.}

\item{...}{Not currently used, but required for extensibility.}

\item{formula}{A formula such as \code{y | x ~ z} specifying the outcome and
predictor terms on the left-hand side, and any covariate terms on the
right-hand side.
The outcome and predictor variables must be separated by a vertical bar
\code{|} on the left-hand side.
This is because EI models the conditional mean of the outcome given
predictors, and how this may vary with covariates.
See the examples for more details.}

\item{data}{When a \strong{formula} is used, \code{data} is a \strong{data frame} containing
both the predictors and the outcome.}

\item{weights}{<\code{\link[rlang:args_data_masking]{data-masking}}> A vector of unit
weights. In general these should be the count of the total number of
individuals in each unit. Required by default. To force constant weights,
you can provide \code{weights=FALSE}.}

\item{bounds}{The bounds to enforce on the outcome, as a length-two vector
\code{c(lower, upper)}. The default \code{NULL} will use the strictest bounds
consistent with the observed data: between 0 and 1, where possible, then
greater than 0, then unbounded.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with numeric columns.
\item A \strong{matrix}
\item A numeric \strong{vector}.
}

When the outcome is a proportion, you can use \code{\link[=ei_proportions]{ei_proportions()}} to assist
in preparing it.
For functions which incorporate bounds on the outcome variable, these are
specified by the \code{bounds} parameter, which is inferred from context, by
default.}

\item{z}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{w} are any covariates,
specified as:
\itemize{
\item A \strong{data frame} with numeric columns.
\item A \strong{matrix}
}

These are scaled internally to have mean zero and unit variance.}
}
\value{
An \code{ei_model} object.
}
\description{
Fits King's (1997) regression model for ecological inference, allowing for overall and
unit-level estimates of conditional means using \code{\link[=ei_est]{ei_est()}}.
}
\examples{
data(elec_1968)

\dontrun{
ei_model(pres_ind_wal ~ vap_white + vap_black, elec_1968, weights=pres_total)
}
}
\references{
King, G. (1997). \emph{A solution to the ecological inference problem:
Reconstructing individual behavior from aggregate data.} Princeton University
Press.
}
