% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_local.R
\name{ei_resid_cov}
\alias{ei_resid_cov}
\title{Estimate the residual covariance of the local estimands}
\usage{
ei_resid_cov(regr, data)
}
\arguments{
\item{regr}{A fitted regression model, from \code{\link[=ei_ridge]{ei_ridge()}}, or another kind
of regression model wrapped with \code{\link[=ei_wrap_model]{ei_wrap_model()}}.}

\item{data}{The data frame, matrix, or \link{ei_spec} object that was used to fit
the regression.}
}
\value{
A covariance matrix. The variables are ordered by predictor within
outcome, e.g. (Y1|X1, Y1|X2, ..., Y2|X1, Y2|X2, ...).
}
\description{
Under a slightly stronger coarsening at random assumption (applying to
second moments), and an assumption of homoskedasticity, this function
estimates the covariance matrix of the local estimands
\eqn{\beta_{gj}=\mathbb{E}[Y | X_j=1, Z=z_g, G=g]} around their local mean.
}
