% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{ei_sens}
\alias{ei_sens}
\title{Conduct a sensitivity analysis for estimated ecological quantities}
\usage{
ei_sens(
  est,
  c_outcome = seq(0, 1, 0.01)^2,
  c_predictor = seq(0, 1, 0.01)^2,
  bias_bound = NULL,
  confounding = 1,
  expand_ci = TRUE
)
}
\arguments{
\item{est}{A set of estimates from \code{\link[=ei_est]{ei_est()}} using both regression and Riesz
representer.}

\item{c_outcome}{The (nonparametric) partial \eqn{R^2} of the omitted
variables with the outcome variables. Must be between 0 and 1.
Can be a vector, in which case all combinations of values with
\code{c_predictor} are used.}

\item{c_predictor}{How much variation latent variables create in the Riesz
representer, i.e. \eqn{1-R^2} of the true Riesz representer on the
estimated one without the omitted variable. Must be between 0 and 1.
Can be a vector, in which case all combinations of values with \code{c_outcome}
are used.}

\item{bias_bound}{If provided, overrides \code{c_predictor} and finds values of
\code{c_predictor} that correspond to (the absolute value of) the provided
amount of bias.}

\item{confounding}{The confounding parameter (\eqn{\rho}), which must be
between 0 and 1 (the adversarial worst-case).}

\item{expand_ci}{If \code{TRUE} and confidence intervals are present in \code{est},
expand the width of the intervals in each direction by the calculated bias
bound.}
}
\value{
A data frame of the same format as \code{est}, but with additional
columns: \code{c_outcome} and \code{c_predictor}, matching all combinations of those
arguments, and \code{bias_bound}, containing the bound on the amount of bias.
The data frame has additional class \code{ei_sens}, which supports a
\code{\link[=plot.ei_sens]{plot.ei_sens()}} method.
}
\description{
Relates confounding of an omitted variable with predictor or outcome to
bias in ecological estimates, using the nonparametric sensitivity analysis
of Chernozhukov et al. (2024).
}
\details{
The parameter \code{c_predictor} equals \eqn{1 - R^2_{\alpha\sim\alpha_s}}, where
\eqn{\alpha} is the true Riesz representer and \eqn{\alpha_s} is the Riesz
representer with the observed covariates. The RR can be equivalently
expressed as \deqn{
  \alpha(n, \bar x_j, z, a) = u(n, \bar x_j)\partial_{\bar x_j} \log f(\bar x_j\mid z, a),
} where \eqn{u(n, \bar x_j)} is the weighting term, \eqn{A} is the unobserved
confounder, and \eqn{f} is the conditional density.
The corresponding \code{c_predictor} is then \deqn{
  1 - R^2_{\alpha\sim\alpha_s} = 1 - \
  \frac{\mathbb{E}[u(N, \bar X_j)(\partial_{\bar x_j} \log f(\bar x_j\mid Z))^2]}{
  \mathbb{E}[u(N, \bar X_j)(\partial_{\bar x_j} \log f(\bar x_j\mid Z, A))^2]}.
}

The bounds here are plug-in estimates and do not incorporate sampling
uncertainty. As such, they may fail to cover the true value in finite
samples, even under large enough sensitivity parameters; see Section 5 of
Chernozhukov et al. (2024).
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(state, pop_urban, farm))
m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)
est = ei_est(m, rr, spec)

ei_sens(est, c_outcome=0.2)

# How much variation would the regression residual need to explain of
# Riesz representer to cause bias of 0.4?
ei_sens(est, c_outcome=1, bias_bound=0.4)

# Update confidence intervals and extract as matrix
est = ei_est(m, rr, spec, conf_level=0.95)
sens = ei_sens(est, c_outcome=0.5, c_predictor=0.2)
as.matrix(sens, "conf.high")

# Works for contrasts as well
est = ei_est(m, rr, spec, contrast = list(predictor=c(1, -1, 0)))
ei_sens(est, c_outcome=0.5, c_predictor=0.5)

}
\references{
McCartan, C., & Kuriwaki, S. (2025+). Identification and semiparametric
estimation of conditional means from aggregate data.
Working paper \href{https://arxiv.org/abs/2509.20194}{arXiv:2509.20194}.

Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2024).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
