% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{ei_sens_rv}
\alias{ei_sens_rv}
\title{Robustness values for ecological inference}
\usage{
ei_sens_rv(est, bias_bound, confounding = 1)
}
\arguments{
\item{est}{A set of estimates from \code{\link[=ei_est]{ei_est()}} using both regression and Riesz
representer.}

\item{bias_bound}{<\code{\link[rlang:args_data_masking]{data-masking}}> A bias bound:
an amount of bias which is considered substantial. Evaluated in the context
of \code{est}, so that one can to refer to \code{std.error} and \code{estimate} as needed.}

\item{confounding}{The confounding parameter (\eqn{\rho}), which must be
between 0 and 1 (the adversarial worst-case).}
}
\value{
A data frame of the same format as \code{est}, but with a new \code{rv} column
containing the robustness values.
}
\description{
The robustness value is the minimum bound for both \code{c_outcome} and
\code{c_predictor} in \code{\link[=ei_sens]{ei_sens()}} such that the bias bound is a certain value.
For example, if the provided bias bound is 0.5, meaning a bias of magnitude
0.5 would be considered problematic, then the robustness value is the minimum
amount of confounding of outcome and predictor (more specifically, the Riesz
representer) that would lead to bias of magnitude 0.5.
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(state, pop_urban, farm))
m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)
est = ei_est(m, rr, spec)

ei_sens_rv(est, 0.1) # how much confounding for bias of 0.1
ei_sens_rv(est, 2 * std.error) # how much confounding for bias of 2 SE

# How much confounding to equalize all estimates (no polarization)
y_avg = weighted.mean(elec_1968$pres_ind_wal, elec_1968$pres_total)
ei_sens_rv(est, estimate - y_avg)

# Extract as matrix
as.matrix(ei_sens_rv(est, 0.2), "rv")

# Works for contrasts as well
est = ei_est(m, rr, spec, contrast = list(predictor=c(1, -1, 0)))
ei_sens_rv(est, estimate) # how much to eliminate disparity

}
\references{
McCartan, C., & Kuriwaki, S. (2025+). Identification and semiparametric
estimation of conditional means from aggregate data.
Working paper \href{https://arxiv.org/abs/2509.20194}{arXiv:2509.20194}.

Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2024).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
