% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{ei_synthetic}
\alias{ei_synthetic}
\title{Generate Synthetic Ecological Data}
\usage{
ei_synthetic(
  x = c(3, 1),
  eta = seq(0.5, 0.9, length.out = ncol(x) \%||\% length(x)),
  Sigma = 0.02 * (1 + diag(ncol(x) \%||\% length(x))),
  n = nrow(x)
)
}
\arguments{
\item{x}{Either a matrix or data frame containing the predictor percentages
in each row, or a vector containing Dirichlet parameters to use in sampling
predictor percentages.}

\item{eta}{The center of the distribution of geography-level parameters.
Defaults to a sequence across groups.}

\item{Sigma}{The covariance matrix for geography-level parameters.}

\item{n}{The number of rows (geographies) to generate. Defaults to the number
of rows in \code{x}, if the latter is a matrix or data frame.}
}
\value{
A list with elements \code{x} (predictors), \code{y} (outcome), \code{b}
(geography-level parameters), and \code{eta} and \code{Sigma} (global parameters).
}
\description{
Samples data from the following truncated Normal ecological model:
\deqn{
     x_i \stackrel{\text{iid}}{\sim} \mathrm{Dirichlet}(\alpha) \\
     b_i \stackrel{\text{iid}}{\sim} \mathcal{N}_{[0, 1]^d}(\eta, \Sigma) \\
     y_i = b_i^\top x_i,
}
where \eqn{\alpha} is a vector of hyperparameters that may be optionally
specified as the \code{x} argument to the function, and \eqn{\mathcal{N}_{[0,
1]^d}} indicates a multivariate Normal distribution truncated to the
\emph{d}-dimensional hypercube.
}
\examples{
ei_synthetic(n=100)

data(elec_1968)
ei_synthetic(cbind(elec_1968$pop_white, 1-elec_1968$pop_white),
     eta=c(0.3, 0.9), Sigma=matrix(c(0.02, 0.016, 0.016, 0.2), nrow=2))

}
