% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_target.R
\name{ei_target}
\alias{ei_target}
\title{Target an ecological regression to respect constraints on the outcome}
\usage{
ei_target(regr, riesz = NULL, data, bounds = NULL, sum_one = NULL)
}
\arguments{
\item{regr}{A fitted regression model, from \code{\link[=ei_ridge]{ei_ridge()}}, or another kind
of regression model wrapped with \code{\link[=ei_wrap_model]{ei_wrap_model()}}.}

\item{riesz}{A fitted Riesz representer from \code{\link[=ei_riesz]{ei_riesz()}}.}

\item{data}{The data frame, matrix, or \link{ei_spec} object used to fit the
regression and Riesz representer.}

\item{bounds}{A vector \code{c(min, max)} of bounds for the outcome. By default,
uses the bounds stored in \code{regr}, if available; otherwise infers bounds
from the outcome variable. Set to \code{FALSE} for unbounded targeting.}

\item{sum_one}{If \code{TRUE}, constrain outcomes to sum to one. By default, uses
the value stored in \code{regr}, if available; otherwise infers the constraint
from the outcome variable and bounds.}
}
\value{
An \code{ei_targeted} object, which is also an \code{ei_wrapped} object and
can be supplied as the \code{regr} argument to \code{\link[=ei_est]{ei_est()}}.
}
\description{
Updates a fitted ecological regression using Targeted Minimum Loss-based
Estimation (TMLE). The returned targeted regression can be supplied to
\code{\link[=ei_est]{ei_est()}} together with the same Riesz representer to produce a targeted
estimate that respects the provided outcome bounds and optional sum-to-one
constraint.
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_abs,
               total = pres_total, covariates = c(pop_urban, farm))
m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)

m_targeted = ei_target(m, rr, spec, bounds = 0:1, sum_one = TRUE)
ei_est(m_targeted, rr, spec)
}
