% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_test_car.R
\name{ei_test_car}
\alias{ei_test_car}
\title{Test the coarsening at random (CAR) assumption}
\usage{
ei_test_car(spec)
}
\arguments{
\item{spec}{An \code{ei_spec} object created with \code{\link[=ei_spec]{ei_spec()}}.}
}
\value{
A 1-row tibble with columns describing the test results. The
}
\description{
The CAR assumption, which is required for accurate estimates with \code{\link[=ei_est]{ei_est()}},
implies that the conditional expectation function (CEF) of the aggregate
outcomes will take a certain partially linear form. This function tests that
implication by comparing a fully nonparametric estimate of the CEF to one
in the partially linear form implied by CAR, and comparing their goodness-of-fit.
It is \strong{very important} that \code{preproc} be used in the \code{spec} to perform a
rich basis transformation of the covariates and predictors; a missing
\code{preproc} will lead to a warning.
}
\details{
The test is carried out by fitting a regression on a fully basis-expanded
combination of covariates and predictors, and calculating the F statistic
compared to the "reduced" model that would be fit by \code{\link[=ei_ridge]{ei_ridge()}}. To
account for penalization, the null distribution is estimated by permuting
the residuals from the reduced model.
}
