% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_wgt.R
\name{ei_wgt}
\alias{ei_wgt}
\alias{ei_wgt_unif}
\alias{ei_wgt_prop}
\alias{ei_wgt_sqrt}
\title{Estimation weighting models}
\usage{
ei_wgt_unif(x)

ei_wgt_prop(x)

ei_wgt_sqrt(x)
}
\arguments{
\item{x}{A numeric vector of unit totals, or an existing \code{\link[=ei_spec]{ei_spec()}} object.}
}
\value{
A numeric vector of estimation weights with the same number of
observations as \code{x}. These will have mean 1.
}
\description{
Several built-in helper functions to generate estimation weights from a
vector of unit totals, or an existing \code{\link[=ei_spec]{ei_spec()}} object.
}
\section{Functions}{
\itemize{
\item \code{ei_wgt_unif()}: Uniform weights across units with any population.
Appropriate if the unit-level variance is constant, i.e., homoskedastic.

\item \code{ei_wgt_prop()}: Weights proportional to the totals. Appropriate if the
unit-level variance is inversely proportional to the number of
observations.

\item \code{ei_wgt_sqrt()}: Weights proportional to the square root of the totals.
Appropriate if the unit-level variance is inversely proportional to the
square root of the number of observations.

}}
\examples{
data(elec_1968)

ei_wgt_unif(head(elec_1968$pres_total))

spec = ei_spec(head(elec_1968), predictors = vap_white:vap_other,
               outcome = pres_ind_wal, total = pres_total)
ei_wgt_prop(spec)
ei_wgt_sqrt(spec)

}
