% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{plot.ei_sens}
\alias{plot.ei_sens}
\title{Bias contour plot for ecological inference estimates}
\usage{
\method{plot}{ei_sens}(
  x,
  y = NULL,
  predictor = NULL,
  bounds = NULL,
  bench = NULL,
  plot_se = 1:3,
  contour_exp = -2:-1,
  ...,
  lwd = 1,
  pch = 8,
  cex = 1
)
}
\arguments{
\item{x}{An \link{ei_sens} object}

\item{y}{An outcome variable, as a character vector. Defaults to first.}

\item{predictor}{A predictor variable to plot, as a character vector. Defaults to first.}

\item{bounds}{A vector \code{c(min, max)} of bounds for the outcome, which will
affect the contours which are plotted. If \code{bounds = NULL} (the default),
they will be inferred where possible. Setting \code{bounds = FALSE} turns off these labels.}

\item{bench}{A data frame of benchmark values, from \code{\link[=ei_bench]{ei_bench()}}, to plot.}

\item{plot_se}{A vector of multiples of the standard error to plot as contours.}

\item{contour_exp}{Powers of 10 for which to plot contours of the bias bound.}

\item{...}{Additional arguments passed on to \code{\link[=contour]{contour()}}}

\item{lwd}{Scaling factor for the contour line widths}

\item{pch}{The point type (see \code{\link[=points]{points()}}) for the benchmark values, if
provided}

\item{cex}{Scaling factor for the benchmark points and labels, if provided}
}
\value{
None, called for side effects (plotting).
}
\description{
Displays bias bound as a function of \code{c_outcome} and \code{c_predictor} in
\code{\link[=ei_sens]{ei_sens()}} on a contour plot. Bounds on the outcome, and standard errors of
the point estimate, can be overlaid as contours on the plot to aid in
interpretation. Benchmarked values of \code{c_outcome} and \code{c_predictor} based on
the observed covariates can also be overlaid.
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(state, pop_urban, farm))
m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)
est = ei_est(m, rr, spec)
sens = ei_sens(est)

plot(sens)

plot(sens, bench = ei_bench(spec), plot_se=NULL)
}
\references{
Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2024).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
