% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_riesz.R
\name{weights.ei_riesz}
\alias{weights.ei_riesz}
\title{Extract Riesz representer weights}
\usage{
\method{weights}{ei_riesz}(object, group = TRUE, loo = FALSE, ...)
}
\arguments{
\item{object}{An \code{\link[=ei_riesz]{ei_riesz()}} object.}

\item{group}{The group for which to extract the weights, as a numeric index
or a character column name. The special (default) value \code{TRUE} will return
a matrix of weights, with each column corresponding to a group.}

\item{loo}{If \code{TRUE}, return the leave-one-out weights}

\item{...}{Additional arguments (ignored)}
}
\value{
A numeric vector of weights
}
\description{
Extracts a single set of Riesz representer weights from an \code{ei_riesz} object,
for a selected group.
}
