% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_ridge.R, R/ei_riesz.R
\name{ei_ridge_impl}
\alias{ei_ridge_impl}
\alias{ei_riesz_impl}
\title{Low-level implementations of \code{ei_ridge()} and \code{ei_riesz()}}
\usage{
ei_ridge_impl(
  x,
  y,
  z,
  weights = rep(1, nrow(x)),
  bounds = c(-Inf, Inf),
  sum_one = NULL,
  penalty = NULL,
  vcov = TRUE
)

ei_riesz_impl(x, z, total, weights = rep(1, nrow(x)), penalty)
}
\arguments{
\item{x}{A matrix of predictors}

\item{y}{A vector of outcomes}

\item{z}{A matrix of covariates}

\item{weights}{A vector of estimation weights}

\item{bounds}{A vector \code{c(min, max)} of bounds for the outcome.}

\item{sum_one}{If \code{TRUE}, the outcome variables are constrained to sum to one.
Can only apply when \code{bounds} are enforced and there is more than one
outcome variable. If \code{NULL}, infers \code{sum_one = TRUE} when the bounds
are \code{c(0, 1)} the outcome variables sum to 1.}

\item{penalty}{The ridge penalty (a non-negative scalar), which must be
specified for \code{\link[=ei_riesz_impl]{ei_riesz_impl()}} but can be automatically estimated with
\code{\link[=ei_ridge_impl]{ei_ridge_impl()}} by providing \code{penalty=NULL}.}

\item{vcov}{If \code{TRUE}, calculate and return the covariance matrix of the
estimated coefficients. When \code{bounds} are provided, the covariance matrix
for the unbounded estimate is returned as a conservative approximation.}

\item{total}{A vector of total observations per unit.}
}
\value{
A list with model components.
}
\description{
No checks are performed on the inputs.
Use of \code{\link[=ei_ridge]{ei_ridge()}} and \code{\link[=ei_riesz]{ei_riesz()}} is strongly recommended unless many
regressions must be fit, e.g., within a tight loop.
Only works for a single outcome, i.e., \code{y} must be a vector, not a matrix.
}
