% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{ei_bench}
\alias{ei_bench}
\title{Benchmark sensitivity parameters from observed covariates}
\usage{
ei_bench(spec, preproc = NULL, subset = NULL)
}
\arguments{
\item{spec}{An \link{ei_spec} object.}

\item{preproc}{An optional function which takes in a data frame of covariates
and returns a transformed data frame or matrix of covariates.
Useful to apply any preprocessing, such as a basis transformation, as part
of the benchmarking process. Passed to \code{\link[rlang:as_function]{rlang::as_function()}}, and so supports
\code{purrr}-style lambda functions.}

\item{subset}{Passed on to \code{\link[=ei_est]{ei_est()}}.}
}
\description{
Produces a table of benchmark values for \code{c_outcome} and \code{c_predictor} in
\code{\link[=ei_sens]{ei_sens()}} for each covariate, following the methodology of Chernozhukov
et al. (2024).
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(educ_elem, pop_urban, farm))

ei_bench(spec)

# preprocess to add all 2-way interactions
ei_bench(spec, preproc = ~ model.matrix(~ .^2 - 1, data = .x))
}
\references{
Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2024).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
