% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{ei_bench}
\alias{ei_bench}
\title{Benchmark sensitivity parameters from observed covariates}
\usage{
ei_bench(spec, preproc = NULL)
}
\arguments{
\item{spec}{An \link{ei_spec} object.}

\item{preproc}{An optional function which takes in a \code{ei_spec} object (\code{spec}
with one covariate removed) and returns a modified object that includes
modified object. Useful to apply any preprocessing, such as a basis
transformation, as part of the benchmarking process.}
}
\description{
Produces a table of benchmark values for \code{c_outcome} and \code{c_predictor} in
\code{\link[=ei_sens]{ei_sens()}} for each covariate, following the methodology of Chernozhukov
et al. (2024).
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(educ_elem, pop_urban, farm))

ei_bench(spec)

# preprocess to add all 2-way interactions
ei_bench(spec, preproc = function(s) {
    z_cols = match(attr(s, "ei_z"), names(s))
    s_out = s[-z_cols]
    z_new = model.matrix(~ .^2 - 1, data = s[z_cols])
    s_out = cbind(s_out, z_new)
    ei_spec(s_out, white:other, pres_ind_wal,
            total = attr(s, "ei_n"), covariates = colnames(z_new))
})
}
\references{
Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2024).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
