% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{ei_bench}
\alias{ei_bench}
\title{Benchmark sensitivity parameters from observed covariates}
\usage{
ei_bench(spec, subset = NULL, contrast = NULL)
}
\arguments{
\item{spec}{An \link{ei_spec} object.}

\item{subset}{<\code{\link[rlang:args_data_masking]{data-masking}}> An optional
indexing vector describing the subset of units over which to calculate
estimates.}

\item{contrast}{If provided, a list containing entries \code{predictor} and
\code{outcome}, each containing a contrast vector.  If only one of \code{predictor}
or \code{outcome} is provided, the contrast will be calculated for all levels of
the other variable.  For example \code{list(predictor = c(1, -1, 0))} will
calculate the difference in each outcome between the first and second
predictor groups; \code{list(outcome = c(1, -1))} will calculate the difference
between the two outcomes for each predictor group; and
\code{list(predictor = c(1, -1, 0), outcome = c(1, -1))} will calculate the
difference in differences.}
}
\description{
Produces a table of benchmark values for \code{c_outcome} and \code{c_predictor} in
\code{\link[=ei_sens]{ei_sens()}} for each covariate, following the methodology of Chernozhukov
et al. (2024).
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(educ_elem, pop_urban, farm))
ei_bench(spec)

# with preprocessed covariates
spec = ei_spec(
    data = elec_1968,
    predictors = vap_white:vap_other,
    outcome = pres_ind_wal,
    total = pres_total,
    covariates = c(educ_elem, pop_urban, farm),
    preproc = ~ model.matrix(~ .^2 - 1, data = .x)
)
ei_bench(spec)
ei_bench(spec, subset = pop_urban > 0.5)

# with contrasts
spec = ei_spec(elec_1968, vap_white:vap_other, pres_rep_nix:pres_ind_wal,
               total = pres_total, covariates = c(educ_elem, pop_urban, farm))
ei_bench(spec, contrast = list(predictor = c(1, -1, 0)))
ei_bench(spec, contrast = list(outcome = c(1, -1)))
}
\references{
Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2024).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
