% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_local.R
\name{ei_local_cov}
\alias{ei_local_cov}
\title{Estimate the residual covariance of the local estimands}
\usage{
ei_local_cov(regr, data, prior_obs = 10)
}
\arguments{
\item{regr}{A fitted regression model, from \code{\link[=ei_ridge]{ei_ridge()}}, or another kind
of regression model wrapped with \code{\link[=ei_wrap_model]{ei_wrap_model()}}.}

\item{data}{The data frame, matrix, or \link{ei_spec} object that was used to fit
the regression.}

\item{prior_obs}{The effective sample size of the inverse-Wishart conjugate prior,
which shrinks the estimate towards the covariance of the regression residuals.
Smaller values mean less shrinkage.}
}
\value{
A covariance matrix. The variables are ordered by predictor within
outcome, e.g. (Y1|X1, Y1|X2, ..., Y2|X1, Y2|X2, ...).
}
\description{
Under a slightly stronger coarsening at random assumption (applying to
second moments), \emph{and} an assumption of homoskedasticity in the covariates,
this function estimates the covariance matrix of the local estimands
\eqn{\beta_{gj}=\mathbb{E}[Y | X_j=1, Z=z_g, G=g]} around their local mean.
See the reference for more detail.
}
\details{
Homoskedasticity in the covariates implies that the variance of the residuals
depends linearly on the entries of \eqn{\bar{X}\bar{X}^\top}. This function
fits an auto-tuned ridge regression of the empirical second moments of the
residuals on these predictors, and uses the polarization identity discussed
in the references to estimate the covariance for each local estimand. When
the estiamated covariance is not positive semidefinite, it is projected onto
the cone of positive semidefinite matrices.
A small amount of shrinkage is applied towards a naive estimator (the
covariance of the regression residuals) under an inverse-Wishart conjugate prior,
whose effective sample size is given by \code{prior_obs}.
}
\references{
McCartan, C., & Kuriwaki, S. (2025+). Identification and semiparametric
estimation of conditional means from aggregate data.
Working paper \href{https://arxiv.org/abs/2509.20194}{arXiv:2509.20194}.
}
