% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\name{ei_proportions}
\alias{ei_proportions}
\title{Convert counts to proportions}
\usage{
ei_proportions(data, ..., .total = ".total", .other = ".other", clamp = 0.001)
}
\arguments{
\item{data}{A data frame.}

\item{...}{<\code{\link[dplyr:select]{tidy-select}}> Columns to convert to proportions.
If one column is completely missing, it will be imputed as 1 minus the total.}

\item{.total}{<\code{\link[dplyr:select]{tidy-select}}> Column to use as the total.
If this column does not exist, it will be created as the sum of the selected
columns. Supports renaming syntax.}

\item{.other}{<\code{\link[dplyr:select]{tidy-select}}> Column to store the remainder,
so that the selected columns plus \code{.other} sum to 1. If the selected
columns do sum to 1 after normalization, this argument will not be used;
otherwise it will be created or overwritten.
The calculation of \code{.other} is performed \emph{after} clamping (see below).}

\item{clamp}{Proportions that are \code{clamp} below 0 or above 1 will be rounded
to 0 and 1, respectively.  Values outside \code{clamp} will throw an error.
Set \code{clamp=0} to disable or \code{clamp=Inf} to allow for out-of-bounds
proportions (not recommended).}
}
\value{
A modified data frame.  Unselected columns are unmodified.
}
\description{
Divides counts in specified columns by a specified total or by their sum,
possibly storing the result in a new column.
Also checks for out-of-bounds and missing values, and can create a column
containing the remainder amount so that all proportions sum to 1.
}
\examples{
data(elec_1968)
# Make a data frame with counts
d_unnorm = with(head(elec_1968, 10), data.frame(
  vap = vap,
  vap_white = vap * vap_white,
  vap_black = vap * vap_black,
  vap_other = vap * vap_other
))

ei_proportions(d_unnorm, vap_white:vap_black, .total=vap) # `.other` column created
ei_proportions(d_unnorm, vap_white:vap_other) # no total provided
# renaming allowed
ei_proportions(d_unnorm, white=vap_white, black=vap_black,
               .total=c(total=vap), .other="vap_other")
}
