% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_ridge.R
\name{ei_ridge}
\alias{ei_ridge}
\alias{ei_ridge.formula}
\alias{ei_ridge.ei_spec}
\alias{ei_ridge.data.frame}
\alias{ei_ridge.matrix}
\alias{ei_ridge.default}
\title{Fit an ecological inference regression model}
\usage{
ei_ridge(x, ...)

\method{ei_ridge}{formula}(formula, data, weights, penalty = NULL, ...)

\method{ei_ridge}{ei_spec}(x, penalty = NULL, ...)

\method{ei_ridge}{data.frame}(x, y, z, weights, penalty = NULL, ...)

\method{ei_ridge}{matrix}(x, y, z, weights, penalty = NULL, ...)

\method{ei_ridge}{default}(x, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item An \link{ei_spec} object containing the outcome, predictor, and covariates.
}

Predictors must be proportions that sum to 1 across rows.
You can use \code{\link[=ei_proportions]{ei_proportions()}} to assist in preparing predictor variables.}

\item{...}{Not currently used, but required for extensibility.}

\item{formula}{A formula such as \code{y | x ~ z} specifying the outcome and
predictor terms on the left-hand side, and any covariate terms on the
right-hand side.
The outcome and predictor variables must be separated by a vertical bar
\code{|} on the left-hand side.
This is because EI models the conditional mean of the outcome given
predictors, and how this may vary with covariates.
See the examples for more details.}

\item{data}{When a \strong{formula} is used, \code{data} is a \strong{data frame} containing
both the predictors and the outcome.}

\item{weights}{<\code{\link[rlang:args_data_masking]{data-masking}}> A vector of unit
weights. In general these should be the count of the total number of
individuals in each unit. Required by default. To force constant weights,
you can provide \code{weights=FALSE}.}

\item{penalty}{The ridge penalty. Set to \code{NULL} to automatically determine
the penalty which minimizes mean-square error, via an efficient
leave-one-out cross validation procedure. The ridge regression solution is
\deqn{\hat\beta = (X^\top X + \lambda I)^{-1}X^\top y,}
where \eqn{\lambda} is the value of \code{penalty}.
Keep in mind when choosing \code{penalty} manually that covariates in \code{z} are
scaled to have mean zero and unit variance before fitting.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with numeric columns.
\item A \strong{matrix}
\item A numeric \strong{vector}.
}

When the outcome is a proportion, you can use \code{\link[=ei_proportions]{ei_proportions()}} to assist
in preparing it.}

\item{z}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{w} are any covariates,
specified as:
\itemize{
\item A \strong{data frame} with numeric columns.
\item A \strong{matrix}
}

These are scaled internally to have mean zero and unit variance.}
}
\value{
An \code{ei_ridge} object, which supports various \link{ridge-methods}.
}
\description{
Fits a penalized regression model for ecological inference, allowing for
overall and unit-level estimates of conditional means using \code{\link[=ei_est]{ei_est()}}.
}
\details{
The regression is calculated using the singular value decomposition, which
allows for efficient recalculation under different \code{penalty} values as part
of leave-one-out cross-validation.
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_oth,
               weights = pres_total, covariates = c(pop_urban, farm))
ei_ridge(spec)

ei_ridge(pres_dem_hum + pres_rep_nix + pres_ind_wal + pres_abs + pres_oth ~
      vap_white + vap_black + vap_other | pop_urban + farm,
    data = elec_1968, weights = pres_total)

}
