% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_ridge.R
\name{ei_ridge}
\alias{ei_ridge}
\alias{ei_ridge.formula}
\alias{ei_ridge.ei_spec}
\alias{ei_ridge.data.frame}
\alias{ei_ridge.matrix}
\alias{ei_ridge.default}
\title{Fit an ecological inference regression model}
\usage{
ei_ridge(
  x,
  ...,
  weights,
  bounds = FALSE,
  sum_one = FALSE,
  penalty = NULL,
  scale = TRUE,
  vcov = TRUE
)

\method{ei_ridge}{formula}(
  formula,
  data,
  weights,
  bounds = FALSE,
  sum_one = FALSE,
  penalty = NULL,
  scale = TRUE,
  vcov = TRUE,
  ...
)

\method{ei_ridge}{ei_spec}(
  x,
  weights,
  bounds = FALSE,
  sum_one = FALSE,
  penalty = NULL,
  scale = TRUE,
  vcov = TRUE,
  ...
)

\method{ei_ridge}{data.frame}(
  x,
  y,
  z,
  weights,
  bounds = FALSE,
  sum_one = FALSE,
  penalty = NULL,
  scale = TRUE,
  vcov = TRUE,
  ...
)

\method{ei_ridge}{matrix}(
  x,
  y,
  z,
  weights,
  bounds = FALSE,
  sum_one = FALSE,
  penalty = NULL,
  scale = TRUE,
  vcov = TRUE,
  ...
)

\method{ei_ridge}{default}(x, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item An \link{ei_spec} object containing the outcome, predictor, and covariates.
}

Predictors must be proportions that sum to 1 across rows.
You can use \code{\link[=ei_proportions]{ei_proportions()}} to assist in preparing predictor variables.
Covariates in an \link{ei_spec} object are shifted to have mean zero. If
\code{scale=TRUE} (the default), they are also scaled to have unit variance.}

\item{...}{Not currently used, but required for extensibility.}

\item{weights}{<\code{\link[rlang:args_data_masking]{data-masking}}> A vector of unit
weights for estimation. These may be the same or different from the total
number of observations in each aggregate unit (see the \code{total} argument to
\code{\link[=ei_spec]{ei_spec()}}). See the discussion below under 'Weights' for choosing this
parameter. The default, uniform weights, makes a slightly
stronger-than-necessary assumption about the relationship between the
unit totals and the unknown data.}

\item{bounds}{A vector \code{c(min, max)} of bounds for the outcome.
If \code{bounds = NULL}, they will be inferred from the outcome variable:
if it is contained within \eqn{[0, 1]}, for instance, then the bounds will
be \code{c(0, 1)}. The default \code{bounds = FALSE} uses an unbounded outcome.}

\item{sum_one}{If \code{TRUE}, the outcome variables are constrained to sum to one.
Can only apply when \code{bounds} are enforced and there is more than one
outcome variable. If \code{NULL}, infers \code{sum_one = TRUE} when the bounds
are \code{c(0, 1)} the outcome variables sum to 1.}

\item{penalty}{The ridge penalty (a non-negative scalar). Set to \code{NULL} to
automatically determine the penalty which minimizes mean-square error,
via an efficient leave-one-out cross validation procedure.
The ridge regression solution is
\deqn{\hat\beta = (X^\top X + \lambda I)^{-1}X^\top y,}
where \eqn{\lambda} is the value of \code{penalty}.
One can equivalently think of the penalty as imposing a
\eqn{\mathcal{N}(0, \sigma^2/\lambda^2)} prior on the \eqn{\beta}.
Keep in mind when choosing \code{penalty} manually that covariates in \code{z} are
scaled to have mean zero and unit variance before fitting.}

\item{scale}{If \code{TRUE}, scale covariates \code{z} to have unit variance.}

\item{vcov}{If \code{TRUE}, calculate and return the a scaled covariance matrix of
the estimated coefficients. When \code{bounds} are provided, the (scaled)
covariance matrix for the unbounded estimate is returned as a conservative
approximation.
The covariance matrix is "scaled" because it does not include the
residual variance. For the covariance for a particular outcome variable,
multiply the returned \verb{$vcov_u} by \code{sigma2} for that outcome.}

\item{formula}{A formula such as \code{y ~ x0 + x1 | z} specifying the outcome \code{y}
regressed on the predictors of interest \code{x} and any covariates \code{z}.
The predictors should form a partition, that is, \code{x0 + x1 = 1} for each
observation. Users can be include more than two predictors as well, e.g.
\code{pct_white + pct_black + pct_hisp + pct_other}.
If there are just two predictors, it is acceptable to only include one in
the formula; the other will be formed as 1 minus the provided predictor.
Include additional covariates separated by a vertical bar \code{|}.
These covariates are strongly recommended for reliable ecological inference.
Covariates are shifted to have mean zero. If \code{scale=TRUE} (the default),
they are also scaled to have unit variance.}

\item{data}{When a \strong{formula} is used, \code{data} is a \strong{data frame} containing
both the predictors and the outcome.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with numeric columns.
\item A \strong{matrix}
\item A numeric \strong{vector}.
}

When the outcome is a proportion, you can use \code{\link[=ei_proportions]{ei_proportions()}} to assist
in preparing it.}

\item{z}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{w} are any covariates,
specified as:
\itemize{
\item A \strong{data frame} with numeric columns.
\item A \strong{matrix}
}

These are shifted to have mean zero. If \code{scale=TRUE} (the default), they
are also scaled to have unit variance.}
}
\value{
An \code{ei_ridge} object, which supports various \link{ridge-methods}.
}
\description{
Fits a penalized regression model for ecological inference, allowing for
overall and unit-level estimates of conditional means using \code{\link[=ei_est]{ei_est()}}.
}
\details{
The regression is calculated using the singular value decomposition, which
allows for efficient recalculation under different \code{penalty} values as part
of leave-one-out cross-validation.
When \code{bounds} are provided, the regression is calculated via quadratic
programming, as there is no closed-form solution. The unbounded regression
is run to select the \code{penalty} automatically in this case, if it is not
provided. Estimation is still efficient, though somewhat slower than in the
unbounded case. The covariance matrix of the estimates is not available when
bounds are applied.
Note that supplying \code{bounds} does not guarantee those bounds will be respected
by \code{\link[=ei_est]{ei_est()}}, and in general it is not necessarily recommended to use \code{bounds}
in \code{\link[=ei_ridge]{ei_ridge()}} any time the outcomes are bounded. Bounds can be enforced
on local estimates in \code{\link[=ei_est_local]{ei_est_local()}}.
}
\section{Weights}{

The weakest identification result for ecological inference makes no
assumption about the number of observations per aggregate unit (the totals).
It requires, however, weighting the estimation steps according to the totals.
This may reduce efficiency when the totals are variable and a slightly
stronger condition holds.

Specifically, if the totals are conditionally mean-independent of the missing
data (the aggregation-unit level means of the outcome within each predictor
level), given covariates, then it is appropriate to use uniform weights in
estimation, or any fixed set of weights.

In general, estimation efficiency is improved when units with larger variance
in the outcome receive less weight. Various bulit-in options are provided by
the helper functions in \code{\link[=ei_wgt]{ei_wgt()}}.
}

\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_abs,
               total = pres_total, covariates = c(pop_urban, farm))
ei_ridge(spec)

ei_ridge(pres_dem_hum + pres_rep_nix + pres_ind_wal + pres_abs ~
      vap_white + vap_black + vap_other | pop_urban + farm, data = elec_1968)

# bounds inferred
all.equal(
  fitted(ei_ridge(spec, bounds = NULL)),
  fitted(ei_ridge(spec, bounds = 0:1))
)

# bounds enforced
min(fitted(ei_ridge(spec)))
min(fitted(ei_ridge(spec, bounds = 0:1)))
}
\references{
McCartan, C., & Kuriwaki, S. (2025+). Identification and semiparametric
estimation of conditional means from aggregate data.
Working paper \href{https://arxiv.org/abs/2509.20194}{arXiv:2509.20194}.
}
