% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_riesz.R
\name{ei_riesz}
\alias{ei_riesz}
\alias{ei_riesz.formula}
\alias{ei_riesz.ei_spec}
\alias{ei_riesz.data.frame}
\alias{ei_riesz.matrix}
\alias{ei_riesz.default}
\title{Estimate Riesz representer for ecological inference}
\usage{
ei_riesz(x, ...)

\method{ei_riesz}{formula}(formula, data, total, weights, penalty, ...)

\method{ei_riesz}{ei_spec}(x, weights, penalty, ...)

\method{ei_riesz}{data.frame}(x, z, total, weights, penalty, ...)

\method{ei_riesz}{matrix}(x, z, total, weights, penalty, ...)

\method{ei_riesz}{default}(x, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item An \link{ei_spec} object containing the outcome, predictor, and covariates.
}

Predictors must be proportions that sum to 1 across rows.
You can use \code{\link[=ei_proportions]{ei_proportions()}} to assist in preparing predictor variables.}

\item{...}{Not currently used, but required for extensibility.}

\item{formula}{A formula such as \code{y | x ~ z} specifying the outcome and
predictor terms on the left-hand side, and any covariate terms on the
right-hand side.
The outcome and predictor variables must be separated by a vertical bar
\code{|} on the left-hand side.
This is because EI models the conditional mean of the outcome given
predictors, and how this may vary with covariates.
See the examples for more details.}

\item{data}{When a \strong{formula} is used, \code{data} is a \strong{data frame} containing
both the predictors and the outcome.}

\item{total}{<\code{\link[tidyselect:language]{tidy-select}}> A variable
containing the total number of observations in each aggregate unit. For
example, the column containing the total number of voters. Required by
default.}

\item{weights}{<\code{\link[rlang:args_data_masking]{data-masking}}> A vector of unit
weights for estimation. These may be the same or different from the total
number of observations in each aggregate unit (see the \code{total} argument to
\code{\link[=ei_spec]{ei_spec()}}). See the discussion below under 'Weights' for choosing this
parameter. The default, uniform weights, makes a slightly
stronger-than-necessary assumption about the relationship between the
unit totals and the unknown data.}

\item{penalty}{The ridge penalty, which must be specified. Recommended value
is the same penalty used in \code{\link[=ei_ridge]{ei_ridge()}}, which is stored in the \code{penalty}
entry of the fitted model object.}

\item{z}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{w} are any covariates,
specified as:
\itemize{
\item A \strong{data frame} with numeric columns.
\item A \strong{matrix}
}

These are scaled internally to have mean zero and unit variance.}
}
\value{
An \code{ei_riesz} object.
}
\description{
Fits a penalized Riesz regression for ecological inference, allowing for
overall estimates of conditional means using \code{\link[=ei_est]{ei_est()}}.
}
\details{
The regression is calculated using the singular value decomposition.
}
\examples{
data(elec_1968)

# Recommended: get ridge penalty from ei_ridge()
spec = ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_oth,
               total = pres_total, covariates = c(pop_urban, farm))
m = ei_ridge(spec)

ei_riesz(spec, penalty = m$penalty)

rr = ei_riesz(~ vap_white + vap_black + vap_other | pop_urban + farm,
              data = elec_1968, total = pres_total, penalty = m$penalty)
summary(rr)

# Examine the weights and check they have mean 1
head(weights(rr, group = "vap_black"))
colMeans(weights(rr))

mean(elec_1968$pres_ind_wal * weights(rr, "vap_white"))
}
