% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{ei_sens}
\alias{ei_sens}
\title{Conduct a sensitivity analysis for estimated ecological quantities}
\usage{
ei_sens(
  regr,
  riesz,
  data,
  subset = NULL,
  c_outcome = seq(0, 1, 0.01)^2,
  c_predictor = seq(0, 1, 0.01)^2,
  bias_bound = NULL,
  confounding = 1,
  conf_level = FALSE
)
}
\arguments{
\item{regr}{A fitted regression model, from \code{\link[=ei_ridge]{ei_ridge()}}, or another kind
of regression model wrapped with \code{\link[=ei_wrap_model]{ei_wrap_model()}}.}

\item{riesz}{A fitted Riesz representer, from \code{\link[=ei_riesz]{ei_riesz()}}, or a matrix of
Riesz weights}

\item{data}{The data frame, matrix, or \link{ei_spec} object that was used to fit
the regression or Riesz representer.}

\item{subset}{<\code{\link[rlang:args_data_masking]{data-masking}}> An optional
indexing vector describing the subset of units over which to calculate
estimates. Not that the subset will not apply to the estimate of
\eqn{\sigma^2\nu^2}, the constant that scales the sensitivity bounds.}

\item{c_outcome}{The (nonparametric) partial \eqn{R^2} of the omitted
variables with the outcome variables. Must be between 0 and 1.
Can be a vector, in which case all combinations of values with
\code{c_predictor} are used.}

\item{c_predictor}{How much variation latent variables create in the Riesz
representer, i.e. \eqn{1-R^2} of the true Riesz representer on the
estimated one without the omitted variable. Must be between 0 and 1.
Can be a vector, in which case all combinations of values with \code{c_outcome}
are used.}

\item{bias_bound}{If provided, overrides \code{c_predictor} and finds values of
\code{c_predictor} that correspond to (the absolute value of) the provided
amount of bias.}

\item{confounding}{The confounding parameter (\eqn{rho}), which must be
between 0 and 1 (the adversarial worst-case).}

\item{conf_level}{A numeric specifying the level for confidence intervals.
If \code{FALSE} (the default), no confidence intervals are calculated. Standard
errors are always returned.}
}
\value{
A data frame of the same format \code{ei_est()}, but with additional
columns: \code{c_outcome} and \code{c_predictor}, matching all combinations of those
arguments, and \code{bias_bound}, containing the bound on the amount of bias.
The data frame has additional class \code{ei_est}, which supports a \code{\link[=plot]{plot()}}
method.
}
\description{
The bounds here are plug-in estimates and do not incorporate sampling
uncertainty. As such, they may fail to cover the true value in finite
samples, even under large enough sensitivity parameters; see Section 5 of
Chernozhukov et al (2022).
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(state, pop_urban, farm))
m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)

ei_sens(m, rr, spec, c_outcome=0.2)

# How much variation would the regression residual need to explain of
# Riesz representer to cause bias of 0.4?
ei_sens(m, rr, spec, c_outcome=1, bias_bound=0.4)

}
\references{
Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2022).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
