% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_est.R
\name{ei_wrap_model}
\alias{ei_wrap_model}
\title{Wrap another predictive model for use in \code{ei_est}}
\usage{
ei_wrap_model(x, data, predictors = NULL, outcome = NULL)
}
\arguments{
\item{x}{A model object, supporting \code{\link[=fitted]{fitted()}} and \code{\link[=predict]{predict()}} generics.}

\item{data}{A data frame or matrix containing the data used to fit the model,
or an \code{\link[=ei_spec]{ei_spec()}} object (recommended). If the latter, then the
\code{predictors} and \code{outcome} arguments are ignored and need not be provided.}

\item{predictors}{<\code{\link[tidyselect:language]{tidy-select}}> Predictor
variables. This is the \code{x} variable in ecological regression that is of
primary interest. For example, the columns containing the percentage of
each racial group.}

\item{outcome}{<\code{\link[tidyselect:language]{tidy-select}}> Outcome
variables. This is the \code{y} variable in ecological regression that is of
primary interest. For example, the columns containing the percentage of
votes for each party.}
}
\value{
An \code{ei_wrapped} object, which has the information required to use
the provided \code{x} with \code{\link[=ei_est]{ei_est()}}.
}
\description{
Stores additional data and attributes on a generic model class so that it
can be used as the \code{regr} argument to \code{\link[=ei_est]{ei_est()}}. Given the wide variety of
model classes, there is no guarantee this function will work. However, most
model classes supporting a \code{\link[=fitted]{fitted()}} and \code{\link[=predict]{predict()}} method will work as long
as there is no transformation of the predictor variables as part of the model
formula or fitting.
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal, pres_total,
               covariates = c(pop_urban, farm))

m = lm(pres_ind_wal ~ 0 + white + black + other + pop_urban + farm, spec)
m_wrap = ei_wrap_model(m, spec)
print(m_wrap)

ei_est(m_wrap, data = spec)

}
