% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_sens.R
\name{plot.ei_sens}
\alias{plot.ei_sens}
\title{Bias contour plot for ecological inference estimates}
\usage{
\method{plot}{ei_sens}(x, y = NULL, predictor = NULL, bounds = NULL, lwd = 1, ...)
}
\arguments{
\item{x}{An \link{ei_sens} object}

\item{y}{An outcome variable, as a character vector. Defaults to first.}

\item{predictor}{A predictor variable to plot, as a character vector. Defaults to first.}

\item{lwd}{A scaling factor for the contour line widths}

\item{...}{Ignored}
}
\description{
TODO fill in...
}
\examples{
data(elec_1968)

spec = ei_spec(elec_1968, vap_white:vap_other, pres_ind_wal,
               total = pres_total, covariates = c(state, pop_urban, farm))
m = ei_ridge(spec)
rr = ei_riesz(spec, penalty = m$penalty)
sens = ei_sens(seq(0, 1, 0.01), seq(0, 1, 0.01), m, rr, spec)

plot(sens)

}
\references{
Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2022).
\emph{Long story short: Omitted variable bias in causal machine learning}
(No. w30302). National Bureau of Economic Research.
}
