% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_form.R
\name{plot.ei_spec}
\alias{plot.ei_spec}
\title{Plot an EI specification}
\usage{
\method{plot}{ei_spec}(x, ..., pch = 16, cex = 0.2)
}
\arguments{
\item{x}{An \link{ei_spec} object.}

\item{...}{Additional arguments passed to \code{\link[=pairs]{pairs()}}.}

\item{pch, cex}{As in \code{\link[=plot]{plot()}}}
}
\description{
Useful for quickly visualizing scatterplots of outcome versus predictor
variables.
}
\examples{
data(elec_1968)
spec = ei_spec(elec_1968, vap_white:vap_other, pres_dem_hum:pres_oth, pres_total)
plot(spec)

}
