% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seine-package.R
\docType{package}
\name{seine-package}
\alias{seine}
\alias{seine-package}
\title{seine: Semiparametric Ecological Inference}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Efficient and user-friendly routines for modern ecological inference. Implements the methods described in McCartan & Kuriwaki (2025+) \doi{10.48550/arXiv.2509.20194}, which generalize ecological regression as introduced by Goodman (1953) \doi{10.2307/2088121}. Includes routines for preprocessing, synthetic data generation, double/debiased machine learning (DML) estimation, partial identification bounds, and sensitivity analysis.
}
\references{
McCartan, C., & Kuriwaki, S. (2025+). Identification and semiparametric
estimation of conditional means from aggregate data.
Working paper \href{https://arxiv.org/abs/2509.20194}{arXiv:2509.20194}.

Kuriwaki, S., & McCartan, C. (2026). The role of confounders and linearity
in ecological inference: A reassessment.
Working paper \href{https://arxiv.org/abs/2601.07668}{arXiv:2601.07668}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corymccartan.com/seine/}
  \item \url{https://github.com/CoryMcCartan/seine}
  \item Report bugs at \url{https://github.com/CoryMcCartan/seine/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID}) [copyright holder]
  \item Shiro Kuriwaki \email{shiro.kuriwaki@yale.edu} (\href{https://orcid.org/0000-0002-5687-2647}{ORCID}) [copyright holder]
}

}
\keyword{internal}
