% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seine-package.R
\docType{package}
\name{seine-package}
\alias{seine}
\alias{seine-package}
\title{seine: Semiparametric Ecological Inference}
\description{
Efficient and user-friendly routines for modern ecological inference. Implements the methods described in McCartan & Kuriwaki (2025+) \href{https://arxiv.org/abs/2509.20194}{arXiv:2509.20194}, which generalize ecological regression as introduced by Goodman (1953) \doi{10.2307/2088121}. Includes routines for preprocessing, synthetic data generation, double/debiased machine learning (DML) estimation, partial identification bounds, and sensitivity analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corymccartan.com/seine/}
  \item \url{https://github.com/CoryMcCartan/seine}
  \item Report bugs at \url{https://github.com/CoryMcCartan/seine/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})

Authors:
\itemize{
  \item Shiro Kuriwaki \email{shiro.kuriwaki@yale.edu} (\href{https://orcid.org/0000-0002-5687-2647}{ORCID})
}

}
\keyword{internal}
