% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pei-package.R
\docType{package}
\name{seine-package}
\alias{seine}
\alias{seine-package}
\title{seine: Semiparametric Ecological Inference}
\description{
Efficient and user-friendly routines for modern ecological inference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CoryMcCartan/seine}
  \item \url{https://corymccartan.com/seine/}
  \item Report bugs at \url{https://github.com/CoryMcCartan/seine/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})

Authors:
\itemize{
  \item Shiro Kuriwaki \email{shiro.kuriwaki@yale.edu} (\href{https://orcid.org/0000-0002-5687-2647}{ORCID})
}

}
\keyword{internal}
